pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Radjin Sardjoe Missier <rmissier@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.1"
uvicorn = "^0.18.2"
pydantic = {extras = ["email"], version = "^1.9.2"}
tortoise-orm = "^0.19.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-dotenv = "^0.20.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
pylint = "^2.14.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
